+++ ui/mods/com.pa.quitch.gwaioverhaul/gw_start/setup.js
@@ -379,7 +379,7 @@ if (!gwoSetupLoaded) {
             var busyToken = {};
             model.makeGameBusy(busyToken);
 
-            var version = "5.55.2";
+            var version = "5.56.0";
             console.log("War created using Galactic War Overhaul v" + version);
 
             var game = new GW.Game();